rule_extractor/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==1.99.4
python-dotenv==1.1.1
python-multipart==0.0.20
PyMuPDF==1.26.3
tiktoken==0.10.0
